# Conflicts:
#	build_scripts/build_usd.py
#	cmake/defaults/gccdefaults.cmake
#	cmake/modules/FindHoudini.cmake
#	extras/imaging/examples/hdTiny/rendererPlugin.cpp
#	pxr/base/arch/stackTrace.cpp
#	pxr/base/plug/initConfig.cpp
#	pxr/base/tf/CMakeLists.txt
#	pxr/base/tf/errorMark.cpp
#	pxr/base/tf/regTest.cpp
#	pxr/imaging/garch/imageRegistry.cpp
#	pxr/imaging/garch/vdbTexture.cpp
#	pxr/imaging/garch/vdbTexture.h
#	pxr/imaging/garch/vdbTextureData.cpp
#	pxr/imaging/garch/vdbTextureData.h
#	pxr/imaging/glf/CMakeLists.txt
#	pxr/imaging/glf/drawTarget.cpp
#	pxr/imaging/glf/package.cpp
#	pxr/imaging/glf/package.h
#	pxr/imaging/glf/plugInfo.json
#	pxr/imaging/glf/shaders/simpleLighting.glslfx
#	pxr/imaging/glf/simpleLightingContext.cpp
#	pxr/imaging/glf/simpleLightingContext.h
#	pxr/imaging/glf/texture.cpp
#	pxr/imaging/glf/utils.cpp
#	pxr/imaging/glf/utils.h
#	pxr/imaging/hdSt/GL/glUtils.cpp
#	pxr/imaging/hdSt/GL/glslProgram.cpp
#	pxr/imaging/hdSt/codeGen.cpp
#	pxr/imaging/hdSt/copyComputation.cpp
#	pxr/imaging/hdSt/drawTarget.h
#	pxr/imaging/hdSt/extCompGpuComputation.cpp
#	pxr/imaging/hdSt/field.cpp
#	pxr/imaging/hdSt/flatNormals.cpp
#	pxr/imaging/hdSt/imageShaderRenderPass.cpp
#	pxr/imaging/hdSt/immediateDrawBatch.cpp
#	pxr/imaging/hdSt/indirectDrawBatch.cpp
#	pxr/imaging/hdSt/instancer.cpp
#	pxr/imaging/hdSt/interleavedMemoryManager.cpp
#	pxr/imaging/hdSt/light.cpp
#	pxr/imaging/hdSt/light.h
#	pxr/imaging/hdSt/quadrangulate.cpp
#	pxr/imaging/hdSt/renderDelegate.h
#	pxr/imaging/hdSt/renderPassShader.cpp
#	pxr/imaging/hdSt/renderPassState.cpp
#	pxr/imaging/hdSt/resourceBinder.cpp
#	pxr/imaging/hdSt/resourceRegistry.cpp
#	pxr/imaging/hdSt/resourceRegistry.h
#	pxr/imaging/hdSt/rprimUtils.cpp
#	pxr/imaging/hdSt/rprimUtils.h
#	pxr/imaging/hdSt/surfaceShader.cpp
#	pxr/imaging/hdSt/unitTestHelper.cpp
#	pxr/imaging/hdSt/vboMemoryManager.cpp
#	pxr/imaging/hdSt/vboMemoryManager.h
#	pxr/imaging/hdx/colorChannelTask.cpp
#	pxr/imaging/hdx/colorChannelTask.h
#	pxr/imaging/hdx/colorCorrectionTask.cpp
#	pxr/imaging/hdx/colorCorrectionTask.h
#	pxr/imaging/hdx/colorizeSelectionTask.cpp
#	pxr/imaging/hdx/colorizeSelectionTask.h
#	pxr/imaging/hdx/colorizeTask.cpp
#	pxr/imaging/hdx/colorizeTask.h
#	pxr/imaging/hdx/fullscreenShader.cpp
#	pxr/imaging/hdx/fullscreenShader.h
#	pxr/imaging/hdx/oitResolveTask.cpp
#	pxr/imaging/hdx/presentTask.cpp
#	pxr/imaging/hdx/presentTask.h
#	pxr/imaging/hdx/shadowTask.cpp
#	pxr/imaging/hdx/simpleLightTask.cpp
#	pxr/imaging/hdx/simpleLightingShader.cpp
#	pxr/imaging/hdx/taskController.cpp
#	pxr/imaging/hdx/taskController.h
#	pxr/imaging/hdx/testenv/testHdxCameraAndLight.cpp
#	pxr/imaging/hdx/testenv/testHdxLightAndShadow.cpp
#	pxr/imaging/hdx/testenv/testHdxRenderTask.cpp
#	pxr/imaging/hgi/buffer.cpp
#	pxr/imaging/hgi/buffer.h
#	pxr/imaging/hgi/hgi.cpp
#	pxr/imaging/hgi/texture.h
#	pxr/imaging/hgi/tokens.h
#	pxr/imaging/hgiGL/buffer.cpp
#	pxr/imaging/hgiGL/buffer.h
#	pxr/imaging/hgiGL/texture.cpp
#	pxr/imaging/hgiGL/texture.h
#	pxr/imaging/hgiMetal/pch.h
#	pxr/imaging/mtlf/package.cpp
#	pxr/imaging/mtlf/shaders/simpleShadowMapShader.glslfx
#	pxr/imaging/mtlf/simpleLight.cpp
#	pxr/imaging/mtlf/utils.h
#	pxr/imaging/plugin/hdEmbree/renderPass.cpp
#	pxr/imaging/plugin/hdEmbree/renderPass.h
#	pxr/usd/sdr/shaderProperty.cpp
#	pxr/usd/usd/schemaRegistry.cpp
#	pxr/usd/usd/schemaRegistry.h
#	pxr/usd/usd/testenv/testUsdSchemaRegistry.py
#	pxr/usd/usd/wrapSchemaRegistry.cpp
#	pxr/usd/usdUtils/sparseValueWriter.h
#	pxr/usdImaging/usdImagingGL/engine.cpp
#	pxr/usdImaging/usdImagingGL/engine.h
#	pxr/usdImaging/usdviewq/appController.py
#	pxr/usdImaging/usdviewq/viewSettingsDataModel.py
#	third_party/houdini/CMakeLists.txt
#	third_party/houdini/gusd/CMakeLists.txt
#	third_party/houdini/gusd/GU_USD.cpp
#	third_party/houdini/gusd/GU_USD.h
#	third_party/houdini/gusd/agentUtils.h
#	third_party/houdini/gusd/shaderWrapper.cpp
#	third_party/houdini/gusd/valueUtils.cpp

+++ pxr/imaging/glf/simpleLightingContext.cpp
@@ -76,8 +76,8 @@ GlfSimpleLightingContext::BindSamplers(GarchBindingMapPtr const &bindingMap)
     
     GLuint _shadowMapTexture = _shadows->GetShadowMapTexture();
     
-    GLuint _shadowMapDepthSampler = (GLuint)(uint64_t)_shadows->GetShadowMapDepthSampler();
-    GLuint _shadowMapCompareSampler = (GLuint)(uint64_t)_shadows->GetShadowMapCompareSampler();
+    GLuint _shadowMapDepthSampler = _shadows->GetShadowMapDepthSampler();
+    GLuint _shadowMapCompareSampler = _shadows->GetShadowMapCompareSampler();
 
     glActiveTexture(GL_TEXTURE0 + shadowSampler);
     glBindTexture(GL_TEXTURE_2D_ARRAY, _shadowMapTexture);
